# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Archiver API",
    "version": "1.0.0",
    "description": "REST API for Open Archiver — an open-source email archiving platform. All authenticated endpoints require a Bearer JWT token obtained from `POST /v1/auth/login`, or an API key passed as a Bearer token.",
    "license": {
      "name": "SEE LICENSE IN LICENSE",
      "url": "https://github.com/LogicLabs-OU/OpenArchiver/blob/main/LICENSE"
    },
    "contact": {
      "name": "Open Archiver",
      "url": "https://openarchiver.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from `POST /v1/auth/login`. Pass as `Authorization: Bearer <token>`."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "API key generated via `POST /v1/api-keys`. Pass as `X-API-KEY: <key>`."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication is required or the token is invalid/expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorMessage"
            },
            "example": {
              "message": "Unauthorized"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated user is not allowed to perform this action.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorMessage"
            },
            "example": {
              "message": "Forbidden"
            }
          }
        }
      },
      "NotFound": {
        "description": "The requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorMessage"
            },
            "example": {
              "message": "Not found"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "An unexpected error occurred on the server.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorMessage"
            },
            "example": {
              "message": "Internal server error"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable error description.",
            "example": "An error occurred."
          }
        },
        "required": [
          "message"
        ]
      },
...
      "Recipient": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true,
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john.doe@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "Attachment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx1y2z3a0000b4d2"
          },
          "filename": {
            "type": "string",
            "example": "invoice.pdf"
          },
          "mimeType": {
            "type": "string",
            "nullable": true,
            "example": "application/pdf"
          },
          "sizeBytes": {
            "type": "integer",
            "example": 204800
          },
          "storagePath": {
            "type": "string",
            "example": "open-archiver/attachments/abc123.pdf"
          }
        },
        "required": [
          "id",
          "filename",
          "sizeBytes",
          "storagePath"
        ]
      },
      "ThreadEmail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ArchivedEmail ID.",
            "example": "clx1y2z3a0000b4d2"
          },
          "subject": {
            "type": "string",
            "nullable": true,
            "example": "Re: Q4 Invoice"
          },
          "sentAt": {
            "type": "string",
            "format": "date-time"
          },
          "senderEmail": {
            "type": "string",
            "format": "email",
            "example": "finance@vendor.com"
          }
        },
        "required": [
          "id",
          "sentAt",
          "senderEmail"
        ]
      },
      "ArchivedEmail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx1y2z3a0000b4d2"
          },
          "ingestionSourceId": {
            "type": "string",
            "example": "clx1y2z3a0000b4d2"
          },
          "userEmail": {
            "type": "string",
            "format": "email",
            "example": "user@company.com"
          },
          "messageIdHeader": {
            "type": "string",
            "nullable": true
          },
          "sentAt": {
            "type": "string",
            "format": "date-time"
          },
          "subject": {
            "type": "string",
            "nullable": true,
            "example": "Q4 Invoice"
          },
          "senderName": {
            "type": "string",
            "nullable": true,
            "example": "Finance Dept"
          },
          "senderEmail": {
            "type": "string",
            "format": "email",
            "example": "finance@vendor.com"
          },
          "recipients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Recipient"
            }
          },
          "storagePath": {
            "type": "string"
          },
          "storageHashSha256": {
            "type": "string",
            "description": "SHA-256 hash of the raw email file, stored at archival time."
          },
          "sizeBytes": {
            "type": "integer"
          },
          "isIndexed": {
            "type": "boolean"
          },
          "hasAttachments": {
            "type": "boolean"
          },
          "isOnLegalHold": {
            "type": "boolean"
          },
          "archivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          },
          "thread": {
            "type": "array",
            "description": "Other emails in the same thread, ordered by sentAt. Only present on single-email GET responses.",
            "items": {
              "$ref": "#/components/schemas/ThreadEmail"
            }
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "required": [
          "id",
          "ingestionSourceId",
          "userEmail",
          "sentAt",
          "senderEmail",
          "recipients",
          "storagePath",
          "storageHashSha256",
          "sizeBytes",
          "isIndexed",
          "hasAttachments",
          "isOnLegalHold",
          "archivedAt"
        ]
      },
      "PaginatedArchivedEmails": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArchivedEmail"
            }
          },
          "total": {
            "type": "integer",
            "example": 1234
          },
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ]
      },
      "SearchResults": {
        "type": "object",
        "properties": {
          "hits": {
            "type": "array",
            "description": "Array of matching archived email objects, potentially with highlighted fields.",
            "items": {
              "type": "object"
            }
          },
          "total": {
            "type": "integer",
            "example": 42
          },
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          },
          "processingTimeMs": {
            "type": "integer",
            "description": "Meilisearch query processing time in milliseconds.",
            "example": 12
          }
        },
        "required": [
          "hits",
          "total",
          "page",
          "limit",
          "totalPages",
          "processingTimeMs"
        ]
      },
...
      "SystemSettings": {
        "type": "object",
        "description": "Non-sensitive system configuration values.",
        "properties": {
          "language": {
            "type": "string",
            "enum": [
              "en",
              "es",
              "fr",
              "de",
              "it",
              "pt",
              "nl",
              "ja",
              "et",
              "el"
            ],
            "example": "en",
            "description": "Default UI language code."
          },
          "theme": {
            "type": "string",
            "enum": [
              "light",
              "dark",
              "system"
            ],
            "example": "system",
            "description": "Default color theme."
          },
          "supportEmail": {
            "type": "string",
            "format": "email",
            "nullable": true,
            "example": "support@example.com",
            "description": "Public-facing support email address."
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/settings/system": {
      "get": {
        "summary": "Get system settings",
        "description": "Returns non-sensitive system settings such as language, timezone, and feature flags. This endpoint is public — no authentication required. Sensitive settings are never exposed.\n",
        "operationId": "getSystemSettings",
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "Current system settings.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettings"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "summary": "Update system settings",
        "description": "Updates system settings. Requires authentication.",
        "operationId": "updateSystemSettings",
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated system settings.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettings"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/search": {
      "get": {
        "summary": "Search archived emails",
        "description": "Performs a full-text search across indexed archived emails using Meilisearch. Requires authentication.",
        "operationId": "searchEmails",
        "tags": [
          "Search"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "keywords",
            "in": "query",
            "required": true,
            "description": "The search query string.",
            "schema": {
              "type": "string",
              "example": "invoice Q4"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number for pagination.",
            "schema": {
              "type": "integer",
              "default": 1,
              "example": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of results per page.",
            "schema": {
              "type": "integer",
              "default": 10,
              "example": 10
            }
          },
          {
            "name": "matchingStrategy",
            "in": "query",
            "required": false,
            "description": "Meilisearch matching strategy. `last` returns results containing at least one keyword; `all` requires all keywords; `frequency` sorts by keyword frequency.",
            "schema": {
              "type": "string",
              "enum": [
                "last",
                "all",
                "frequency"
              ],
              "default": "last"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResults"
                }
              }
            }
          },
          "400": {
            "description": "Keywords parameter is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/v1/archived-emails/ingestion-source/{ingestionSourceId}": {
      "get": {
        "summary": "List archived emails for an ingestion source",
        "description": "Returns a paginated list of archived emails belonging to the specified ingestion source. Requires authentication.",
        "operationId": "getArchivedEmails",
        "tags": [
          "Archived Emails"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "ingestionSourceId",
            "in": "path",
            "required": true,
            "description": "The ID of the ingestion source to retrieve emails for.",
            "schema": {
              "type": "string",
              "example": "clx1y2z3a0000b4d2"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number for pagination.",
            "schema": {
              "type": "integer",
              "default": 1,
              "example": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of items per page.",
            "schema": {
              "type": "integer",
              "default": 10,
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of archived emails.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedArchivedEmails"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/archived-emails/{id}": {
      "get": {
        "summary": "Get a single archived email",
        "description": "Retrieves the full details of a single archived email by ID, including attachments and thread. Requires authentication.",
        "operationId": "getArchivedEmailById",
        "tags": [
          "Archived Emails"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the archived email.",
            "schema": {
              "type": "string",
              "example": "clx1y2z3a0000b4d2"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archived email details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchivedEmail"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete an archived email",
        "description": "Permanently deletes an archived email by ID. Deletion must be enabled in system settings and the email must not be on legal hold. Requires authentication.",
        "operationId": "deleteArchivedEmail",
        "tags": [
          "Archived Emails"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the archived email to delete.",
            "schema": {
              "type": "string",
              "example": "clx1y2z3a0000b4d2"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Email deleted successfully. No content returned."
          },
          "400": {
            "description": "Deletion is disabled in system settings, or the email is blocked by a retention policy / legal hold.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
...
}